requirements.txt
boto3==1.11.3
botocore==1.14.3
Click==7.0
colorzero==1.1
docutils==0.15.2
Flask==1.1.1
gpiozero==1.5.1
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
python-dateutil==2.8.1
RPi.GPIO==0.7.0
s3transfer==0.3.0
six==1.14.0
urllib3==1.25.7
Werkzeug==0.16.0
